+++ core/src/test/java/nl/alexeyu/structmatcher/matcher/ValueBox.java
@@ -1,0 +1,8 @@
+package nl.alexeyu.structmatcher.matcher;
+
+/** A generic accessor interface; {@link StringValueBox} implements it covariantly. */
+public interface ValueBox<T> {
+
+    T getValue();
+
+}